NAS-121735: Move questions to separate file
